requirements.txt
opencv-python >= 4.2.0.34
numpy >= 1.18.4
pillow >= 7.1.2 